// frontend/src/app/api/models/resultado-busqueda.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImagenPublicaNula } from '../models/imagen-publica-nula';
import { Slug } from '../models/slug';
export interface ResultadoBusqueda {
  pais?: (string | null);
  portada?: ImagenPublicaNula;
  resumen?: (string | null);
  slug: Slug;
  tipo: 'DESTINO' | 'ITINERARIO' | 'GUIA' | 'TIPO';
  titulo: string;
}